Add files via upload
The study adoptes a  full hierarchical Bayesian models to estimate the relative risk of the occurrence of TB via Integrated Nested Laplace (INLA) in R. The comparison is focused on how well each of these methods identify the regions which have high risks. The best specified model included all the selected covariates and both the spatially and non-spatially structured random effects (GLMM). The gain in goodness-of-fit statistic (DIC) when the spatial structure was included confirms the clustered spatial pattern of disease and GLMM as the best model.

Rstudio Version 1.2.5019 was used in implementing the code.